SystemUI: fix navbar dpad arrow keys regression
A previous commit (3dcd3654257f0cf41b50ad3a544eaa7902386a4a) which
was supposed to improve the dpad showing logic when swiching IMEs caused
a regression in that it did not hide small navbar buttons on the side
which were not the menu button. This is because the logic to store the
previous butotn visibility went through setButtonWithTagVisibility()
which is really only called on the primary buttons + menu button, and
never the search button.

Always store the previous visibilities of the buttons we're about to
override, so they can be restored properly.

Change-Id: I6d0dee7fe7918613b5d8ca7a81adfd631acffc1e
Signed-off-by: Roman Birg <roman@cyngn.com>