SysytemUI: Show the operator name in keyguard and notify bar
Add a flag to check if the operator label in Keyguard is needed.
Move the operator name from status bar to notify bar to make sure
it could be shown even when there is any notify.

By default, it shows operator name in Keyguard but not in notify bar
to follow the Android original design.
Customize to show/hide operator name label in overlay of data package
if needed.

Change-Id: Ieb540f25c384fafa840c06c2691e91506e9ef5a6