OMS: Allow overlays to move to state 2 after a delay of 10 seconds
An attempt to let overlays hold their activation values for 10 seconds
while an app is updating. If an app is uninstalled for more than 10
seconds, then remove the overlay, however if the app is updated and the
uninstall phase was under 10 seconds, then put the overlay back to it’s
original state.